refactor(discord): consolidate member commands into single MemberCommand
Consolidates register, create-member, and approve-member commands
into a single /member command with subcommands:
- /member create <name> [github]: Register as a member
- /member approve <org> <user> <action>: Approve/reject organization join

Removes duplicate register/create-member commands and groups related
member management functionality.
